SUCCESFULLY CREATED WEB SERVER
Managded to connect client to the web server. The LittleFS lib allows the user to connect to web server.
Needed to upload code to flash and to start web server, but it works just fine with every page.

Also web client html css and script files were moved to data folder in order to connect to them with the ESP32.
Connection only good if esp32 and client are connected to the same network